commit #11.02

+++ src/main/java/ru/customer/service/SockServiceImpl.java
@@ -1,13 +1,22 @@
-package ru.job4j.service;
+package ru.customer.service;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
-import ru.job4j.model.Sock;
-import ru.job4j.model.SockResponse;
-import ru.job4j.repository.SocksRepository;
+import ru.customer.model.Color;
+import ru.customer.model.Sock;
+import ru.customer.model.SockResponse;
+import ru.customer.repository.ColorRepositoryImpl;
+import ru.customer.repository.SocksRepository;
+import ru.customer.utils.AppConstants;
 
 import javax.annotation.PostConstruct;
 import java.util.List;
@@ -21,45 +30,32 @@ public class SockServiceImpl implements SockService {
 
     static final Logger LOGGER = LoggerFactory.getLogger(SockServiceImpl.class);
 
-    private final ConcurrentHashMap<String, String> colors = new ConcurrentHashMap<>();
     private final ConcurrentHashMap<String, String> operators = new ConcurrentHashMap<>();
 
     private final SocksRepository repository;
+    private final ColorRepositoryImpl colorRepository;
 
     @PostConstruct
     private void init() {
         initialization();
     }
 
-//todo перенести в БД
     void initialization() {
-        colors.put("white", "white");
-        colors.put("black", "black");
-        colors.put("red", "red");
-        colors.put("blue", "blue");
-        colors.put("green", "green");
-        colors.put("yellow", "yellow");
-        colors.put("orange", "orange");
-        colors.put("gray", "gray");
         operators.put("moreThan", "moreThan");
         operators.put("lessThan", "lessThan");
         operators.put("equal", "equal");
     }
 
-    public String matchesColor(String string) {
-        String rsl = "not registered";
-        if (colors.contains(string)) {
-            return colors.get(string);
-        }
-        return rsl;
+    public String matchesColor(String coloring) {
+        return colorRepository.existsColorByColoring(coloring)
+                ? colorRepository.findColorByColoring(coloring).get().getColoring()
+                : AppConstants.RESPONSE;
     }
 
     public String matchesOperator(String operator) {
-        String rsl = "not registered";
-        if (operators.contains(operator)) {
-            return operators.get(operator);
-        }
-        return rsl;
+        return operators.contains(operator)
+                ? operators.get(operator)
+                : AppConstants.RESPONSE;
     }
 
     public Iterable<Sock> findAll() {
@@ -67,7 +63,7 @@ public class SockServiceImpl implements SockService {
     }
 
     /**
-     * //todo color
+     * //todo color - test logger
      * Найти по параметру цвет и % хлопка и увеличить на кол-во единиц Quantity,
      * Сохранить сущность в БД согласно указанных параметров
      *
@@ -75,19 +71,29 @@ public class SockServiceImpl implements SockService {
      * @return сохраненная сущность Sock c увеличенным числом пар
      */
     public Sock save(Sock sock) {
-        var rsl = repository
-                .findByColorAndCottonPartWithin(sock.getColor(), sock.getCottonPart());
-        if (rsl.isPresent()) {
-            var tempSock = rsl.get();
-            var count = tempSock.getQuantity() + sock.getQuantity();
-            tempSock.setQuantity(count);
-            return repository.save(tempSock);
+        var color = findByColor(sock.getColor().getColoring());
+        if (color.isPresent()) {
+            var rsl = repository
+//                    .findByColorAndCottonPartWithin(
+//                            color.get().getId(), sock.getCottonPart());
+                    .findSockByColorAndCottonPart(color.get(), sock.getCottonPart());
+            LOGGER.info("Что нашли -> {}", rsl);
+            if (rsl.isPresent()) {
+                var tempSock = rsl.get();
+                tempSock.setQuantity(tempSock.getQuantity() + sock.getQuantity());
+                return repository.save(tempSock);
+            }
         }
+        sock.setColor(colorRepository.save(sock.getColor()));
         return repository.save(sock);
     }
 
+    private Optional<Color> findByColor(String coloring) {
+        return colorRepository.findColorByColoring(coloring);
+    }
+
     /**
-     * //todo color Object
+     * //todo color Object - test logger
      * Найти по параметрам color and cottonPart
      * и уменьшить количество единиц товара(Quantity) согласно заданного параметра
      *
@@ -96,39 +102,33 @@ public class SockServiceImpl implements SockService {
      */
     public Sock reduceSockQuantity(Sock sock) {
         var rsl = repository
-                .findByColorAndCottonPartWithin(sock.getColor().getColoring(),
+                .findSockByColorAndCottonPart(sock.getColor(),
                         sock.getCottonPart());
+        return findAndDeleteSock(rsl, sock);
+    }
+
+    private Sock findAndDeleteSock(Optional<Sock> rsl, Sock sock) {
         if (rsl.isPresent()) {
             var tempSock = rsl.get();
-            var count = tempSock.getQuantity() - sock.getQuantity();
-            tempSock.setQuantity(count);
-            if (tempSock.getQuantity() > 0) {
-                return repository.save(tempSock);
+            if (tempSock.getQuantity() > sock.getQuantity()) {
+                tempSock.setQuantity(tempSock.getQuantity() - sock.getQuantity());
+                if (tempSock.getQuantity() > 0) {
+                    return repository.save(tempSock);
+                }
+                repository.delete(tempSock);
             }
-            repository.delete(tempSock);
         }
         return Sock.of(null, 0, 0);
     }
 
     /**
-     * найти по ID
+     * Find by ID
      *
-     * @param id int
+     * @param id Long
      * @return Optional<Sock>
      */
-    public Optional<Sock> findById(int id) {
+    public Optional<Sock> findById(Long id) {
         return repository.findById(id);
-    }
-
-    /**
-     * найти согласно цвета сущности
-     * color — цвет носков, строка (например, black, red, yellow);
-     *
-     * @param color
-     * @return List<Sock>
-     */
-    public List<Sock> findByColor(String color) {
-        return repository.findAllByColor(color);
     }
 
     /**
@@ -140,37 +140,6 @@ public class SockServiceImpl implements SockService {
      */
     public List<Sock> findByCottonPart(int cottonPart) {
         return repository.findAllByCottonPart(cottonPart);
-    }
-
-    /**
-     * получить все пару носков где color and cottonPart совпадают заданных параметрам
-     *
-     * @param color
-     * @param cottonPath
-     * @return Sock Entity
-     */
-    public Sock findByColorAndCottonPart(String color, int cottonPath) {
-        var rsl = repository.findByColorAndCottonPartWithin(color, cottonPath);
-        if (rsl.isPresent()) {
-            return rsl.get();
-        }
-        return Sock.of(null, 0, 0);
-    }
-
-    /**
-     * //todo color Object
-     * найти по 2-м параметрам и уменьшить общее колл-во на 3-й параметр
-     *
-     * @param sock запрос с объектом
-     * @return sock итоговый остаток объектов в БД
-     */
-    public Sock reducesByQuantity(Sock sock) {
-        var rsl = findByColorAndCottonPart(sock.getColor(), sock.getCottonPart());
-        LOGGER.info("what find {}", rsl);
-        var count = rsl.getQuantity() - sock.getQuantity();
-        rsl.setQuantity(count);
-        repository.save(rsl);
-        return rsl;
     }
 
     /**
@@ -187,15 +156,30 @@ public class SockServiceImpl implements SockService {
      * @return List<Sock> Entity
      */
     public List<Sock> findByOperator(String color, String operator, int cottonPart) {
-        if (operator.equals("moreThan")) {
-            return repository.findAllByColorAndCottonPartAndMore(color, cottonPart);
+        var rsl = colorRepository.findColorByColoring(color);
+        LOGGER.info("Что нашлось в репозитории цвета -> " + rsl);
+        if (rsl.isPresent()) {
+            var rslColor = rsl.get();
+            if (operator.equals(AppConstants.MORE_THAN)) {
+                return repository.findSocksByColorAndCottonPartIsGreaterThanOrderById(rslColor, cottonPart)
+                        .orElse(List.of(Sock.of(null, 0, 0)));
+            }
+            if (operator.equals(AppConstants.LESS_THAN)) {
+                return repository.findSocksByColorAndCottonPartAndSmaller(rslColor, cottonPart)
+                        .orElse(List.of(Sock.of(null, 0, 0)));
+            }
+
+            return List.of(repository.findSockByColorAndCottonPart(rslColor, cottonPart)
+                    .orElse(Sock.of(null, 0, 0)));
         }
-        if (operator.equals("lessThan")) {
-            return repository.findAllByColorAndCottonPartAndSmaller(color, cottonPart);
-        }
-        return List.of(findByColorAndCottonPart(color, cottonPart));
+        return List.of(Sock.of(null, 0, 0));
     }
 
+    /**
+     * Find the id of the last saved entity
+     *
+     * @return Long id entity
+     */
     public Long findIdLastEntity() {
         var rsl = repository.findFirstByOrderByIdDesc();
         return rsl.isPresent() ? rsl.get().getId() : 0L;
@@ -212,7 +196,21 @@ public class SockServiceImpl implements SockService {
      */
     @Override
     public SockResponse getAllSock(int pageNo, int pageSize, String sortBy, String sortDir) {
-        return null;
+        Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending()
+                : Sort.by(sortBy).descending();
+        Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
+        Page<Sock> socks = (Page<Sock>) repository.findAll((Specification<Sock>) pageable);
+        //todo Page<Sock> socks = repository.findAll(pageable);
+        List<Sock> listOfSock = socks.getContent();
+
+        SockResponse carResponse = new SockResponse();
+        carResponse.setContent(listOfSock);
+        carResponse.setPageNo(socks.getNumber());
+        carResponse.setPageSize(socks.getSize());
+        carResponse.setTotalElements(socks.getTotalElements());
+        carResponse.setTotalPages(socks.getTotalPages());
+        carResponse.setLast(socks.isLast());
+        return carResponse;
     }
 
     public boolean deleteById(Long id) {
